Modify structure for wallet and cli utilization
