+++ wails/wallet/cmd/cli/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"fmt"
+	"os"
+)
+
+func main() {
+	fmt.Println("args: ", os.Args[1:])
+}
