+++ wails/wallet/internal/currencies/eth/hardhat.go
@@ -1,4 +1,4 @@
-package currencies
+package eth
 
 import (
 	"bytes"
